+++ bookstore/api/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-
-from bookstore import settings
-from .views import AuthorViewSet, BookViewSet
+from django.conf import settings
+from api.viewsets import AuthorViewSet, BookViewSet
 
 router = DefaultRouter()
 router.register(r'authors', AuthorViewSet)
