+++ bookstore/api/viewsets/__init__.py
@@ -1,0 +1,2 @@
+from .author_viewset import AuthorViewSet
+from .book_viewset import BookViewSet
