+++ bookstore/api/viewsets/author_viewset.py
@@ -1,0 +1,7 @@
+from rest_framework import viewsets
+from ..models import Author
+from ..serializers import AuthorSerializer
+
+class AuthorViewSet(viewsets.ModelViewSet):
+    queryset = Author.objects.all()
+    serializers_class = AuthorSerializer
