+++ bookstore/api/viewsets/book_viewset.py
@@ -1,0 +1,7 @@
+from rest_framework import viewsets
+from ..models import Book
+from ..serializers import BookSerializer
+
+class BookViewSet(viewsets.ModelViewSet):
+    queryset = Book.objects.all()
+    serializer_class = BookSerializer
